pyproject.toml
[project]
name = "recharge-connector"
version = "1.2.0"
description = "Utility to retrieve data from Recharge"
readme = "README.md"
authors = [{ name = "Dan Ross", email = "daniel.ross@aop.com" }]
requires-python = ">=3.12"
dependencies = [
    "fastexcel>=0.14.0",
    "polars>=1.27.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "tqdm>=4.67.1",
    "xlsxwriter>=3.2.3",
]

[project.scripts]
recharge-connector = "recharge_connector:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
